py: rename 'aux' folder to 'auxiliary'
Cloning the xnvme repository on Windows fails with the message "invalid
path". The reason is that "aux" is a protected name in NTFS. This change
renames the "aux" folder to "auxiliary" to work around this issue.
